[TIMOB-24332] Move Kroll-Core to ARC
