+++ iphone/Classes/KrollPropertyDelegate.m
@@ -14,16 +14,10 @@
 {
 	if (self = [super init])
 	{
-		target = [target_ retain];
+        target = target_;
 		selector = selector_;
 	}
 	return self;
-}
-
--(void)dealloc
-{
-	[target release];
-	[super dealloc];
 }
 
 -(id)target
