+++ iphone/Classes/TiDOMCharacterDataProxy.h
