+++ iphone/Classes/TiDOMCharacterDataProxy.m
