+++ iphone/Classes/defines.h
@@ -7,6 +7,7 @@
 
 // MODULES
 
+#define TI_USE_KROLL_THREAD
 #define USE_TI_STREAM
 #define USE_TI_CODEC
 #define USE_TI_UTILS
